+++ jerry-core/api/jerry-snapshot.c
