+++ jerry-core/api/jerry-snapshot.h
