+++ jerry-core/api/jerry.c
