+++ targets/mbedos5/jerryscript-mbed/jerryscript-mbed-library-registry/wrap_tools.h
@@ -17,8 +17,8 @@
 
 #include <stdlib.h>
 
-#include "jerry-core/jerryscript.h"
- 
+#include "jerry-core/include/jerryscript.h"
+
 #include "jerryscript-mbed-util/logging.h"
 #include "jerryscript-mbed-util/wrappers.h"
