+++ targets/mbedos5/source/jerry_port_mbed.c
@@ -18,7 +18,7 @@
 #include <stdlib.h>
 #include <sys/time.h>
 
-#include "jerry-core/jerryscript-port.h"
+#include "jerry-core/include/jerryscript-port.h"
 
 #include "us_ticker_api.h"
